refactor for login component
