+++ app/(private)/(tabs)/_layout.tsx
@@ -2,10 +2,10 @@ import { Tabs } from 'expo-router';
 import React from 'react';
 import { Platform } from 'react-native';
 
-import { HapticTab } from '@/components/HapticTab';
-import { IconSymbol } from '@/components/ui/IconSymbol';
-import TabBarBackground from '@/components/ui/TabBarBackground';
-import { Colors } from '@/constants/Colors';
+import { HapticTab } from '@/components/ui/interactions/HapticTab';
+import { IconSymbol } from '@/components/ui/icons/IconSymbol';
+import TabBarBackground from '@/components/tabs/TabBarBackground';
+import { Colors } from '@/components/ui/themes/Colors';
 import { useColorScheme } from '@/hooks/useColorScheme';
 
 export default function TabLayout() {
