+++ app/(private)/_layout.tsx
@@ -3,12 +3,9 @@ import { Stack } from 'expo-router';
 export default function ProtectedLayout() {
     return (
         <ProtectedRoute>
-            <Stack>
+            <Stack screenOptions={{ headerShown: false }}>
                 <Stack.Screen name="(tabs)" options={{ headerShown: false }} />
-                <Stack.Screen
-                    name="ChatRoom"
-                    options={{ headerShown: true }}
-                />
+                <Stack.Screen name="chat-room" options={{ headerShown: true }} />
             </Stack>
         </ProtectedRoute>
     )
