+++ app/(public)/login.tsx
@@ -3,43 +3,19 @@ import { StyleSheet, FlatList, SafeAreaView } from 'react-native';
 import { useRouter } from 'expo-router';
 import { StatusBar } from 'expo-status-bar';
 import { useAppContext } from '@/hooks/AppContext';
-import { ThemedText } from '@/components/ThemedText';
-import { ThemedView } from '@/components/ThemedView';
-import { UserListItem } from '@/components/UserListItem';
+import { ThemedText } from '@/components/ui/text/ThemedText';
+import { ThemedView } from '@/components/ui/layout/ThemedView';
+import { UserListItem } from '@/components/users/UserListItem';
+import LoginList from '@/components/login/LoginList';
 
 export default function LoginScreen() {
   const { users, login } = useAppContext();
   const router = useRouter();
 
-  const handleUserSelect = async (userId: string) => {
-    if (await login(userId)) {
-      router.replace('/(private)/(tabs)');
-    }
-  };
-
   return (
     <SafeAreaView style={styles.safeArea}>
       <StatusBar style="auto" />
-      <ThemedView style={styles.container}>
-        <ThemedView style={styles.header}>
-          <ThemedText type="title">Welcome to Chat App</ThemedText>
-          <ThemedText style={styles.subtitle}>
-            Select a user to continue
-          </ThemedText>
-        </ThemedView>
-        
-        <FlatList
-          data={users}
-          keyExtractor={(item) => item.id}
-          renderItem={({ item }) => (
-            <UserListItem
-              user={item}
-              onSelect={() => handleUserSelect(item.id)}
-            />
-          )}
-          contentContainerStyle={styles.listContainer}
-        />
-      </ThemedView>
+      <LoginList></LoginList>
     </SafeAreaView>
   );
 }
@@ -47,22 +23,5 @@ export default function LoginScreen() {
 const styles = StyleSheet.create({
   safeArea: {
     flex: 1,
-  },
-  container: {
-    flex: 1,
-    paddingTop: 20,
-  },
-  header: {
-    alignItems: 'center',
-    padding: 20,
-    marginBottom: 20,
-  },
-  subtitle: {
-    marginTop: 10,
-    fontSize: 16,
-    color: '#8F8F8F',
-  },
-  listContainer: {
-    paddingBottom: 20,
   },
 });
