+++ components/tabs/TabBarBackground.tsx
