+++ components/ui/icons/IconSymbol.ios.tsx
