+++ components/ui/interactions/HapticTab.tsx
